[face-detector] Make firebase pod version overridable
This makes the firebase pod version overridable so that this package can be used in bare projects, together with other firebase packages. When Firebase versions differ accross multiple packages, pod install will fail. This follows the `$FirebaseSDKVersion` constants also used by react-native-firebase to maximize interop and make it easier for users to configure.